Add files via upload
Support Vector Machine (SVM) is an algorithm that can be implemented in either classification or regression problems. SVM is a supervised machine learning algorithm. In our project we aim to implement the SVM algorithm on the MNIST dataset, handwritten digits dataset, to classify digits using scikit-learn. A 3-fold cross-validation and hyperparameter selection study is going to be conducted to find out the best parameters for the MNIST dataset classification. Three kernel functions namely rbf, poly, and sigmoid are assumed. In order to evaluate the performance of SVM for each kernel function, the accuracy, precision, and recall for each fold are reported. Finally, confusion matrix and aforementioned parameters are also reported.